requirements.txt
flask
gunicorn
azure-identity
azure-keyvault-secrets
azure-storage-blob